+++ R/01-getMappingTable.R
@@ -1,9 +1,15 @@
 #' getMappingTable
 #'
+#' @inheritParams createNewFileSource
 #' @export
-getMappingTable <- function() {
-  mappingfile <- system.file("mapping", "Field_Mapping.csv", package = "MpiIsoData")
-  mapping <- read.csv2(file = mappingfile, stringsAsFactors = FALSE, check.names = FALSE,
+getMappingTable <- function(mappingName) {
+  if (!(paste0(mappingName, ".csv") %in% dir(system.file('mapping', package = 'MpiIsoData')))) {
+    stop("Mapping not found! Please add the mapping file to inst/mapping/.")
+  }
+
+  mappingFile <- paste0(mappingName, ".csv")
+  mappingFile <- system.file("mapping", mappingFile, package = "MpiIsoData")
+  mapping <- read.csv2(file = mappingFile, stringsAsFactors = FALSE, check.names = FALSE,
                        na.strings = c("", "NA"), strip.white = TRUE)
   mapping
 }