+++ inst/sql/removeOldSourceFromData.sql
@@ -1,0 +1,1 @@
+DELETE FROM `{{ mappingId }}_data` where `source` not in {{ sources }};