+++ inst/sql/removeOldSourceFromTable.sql
@@ -1,0 +1,1 @@
+DELETE FROM `{{ table }}` where `mappingId` = '{{ mappingId }}' AND `source` not in {{ sources }};